Fix TypeError when elevated_session_duration is a string
Cast the config value to int before passing to Carbon's addMinutes(),
which requires int|float. This fixes the issue when users configure
the value via env() which returns strings.

Fixes #14769

+++ src/Providers/AppServiceProvider.php
@@ -357,7 +357,7 @@ class AppServiceProvider extends ServiceProvider
             }
 
             return Carbon::createFromTimestamp($lastElevated)
-                ->addMinutes(config('statamic.users.elevated_session_duration', 15))
+                ->addMinutes((int) config('statamic.users.elevated_session_duration', 15))
                 ->timestamp;
         });
